+++ hermes_legion_commander/__init__.py
@@ -1,3 +1,3 @@
 """Hermes Legion Commander package."""
 
-__version__ = "1.7.2"
+__version__ = "1.7.4"